refactor: move derive_seed to utils.seeds (closes #52)
derive_seed is a general-purpose blake2b-based stable-seed helper, not
task-sampling-specific.  Moves it alongside utils.codes so other modules
that need cross-process-stable seeding can import it without pulling in
sample_tasks.

Import path change: from every_query.utils.seeds import derive_seed.
Also expanded the docstring to explain why the 0x7FFFFFFF mask is the
right return-value cap (NumPy's legal seed range).
